Address CR comments
